requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.0.12
cloudinary==1.29.0
confusable-homoglyphs==3.2.0
Django==3.2.10
django-bootstrap3==21.2
django-bootstrap4==22.1
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-registration==2.4.1
django-utils-six==2.0
djangorestframework==3.13.1
idna==3.3
Pillow==9.1.1
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
